oby4-sd/oby4-wf: revise downstream device identifier and response This commit revise the content of the downstream device identifier and response, it will response with out retimer's attribute if they're not present in the instance; it a;so replace with different model string for VR to distinguish different VR source. 1. The original descriptor doesn't distinguish different source of component and may mislead end user to flash it with the wrong firmware.
2. Current implementation remain exposing attribute of Retimer even
they're not present in the machine.

3. Current format doesn't meet the requirement from enduser.
1. The `ComponentType` and `ComponentInstance` attribute is used to
expose on D-Bus by PLDM, assume it is valid if PLDM expose it properly.

2. `PLDM_ASCII_MODEL_NUMBER_LONG_STRING` and
`PLDM_ASCII_MODEL_NUMBER_SHORT_STRING` attribute are used to identify
all unique devices in the instance, and use to filter the PLDM Package
to prevent device be flashed by wrong firmware, in the test, the right
and wrong firmware package will be provided to test if PLDM can
distinguished by this.

> Conflicts:
>	meta-facebook/yv4-sd/src/platform/plat_pldm_fw_update.c